+++ coloraide/spaces/zcam.py
@@ -252,9 +252,15 @@ def zcam_to_xyz(
     if env is None:
         raise ValueError("No viewing conditions/environment provided")
 
-    # Black
-    if Jz == 0.0 or Qz == 0.0:
-        return [0.0, 0.0, 0.0]
+    # Black?
+    if Jz == 0.0:
+        Jz = alg.EPS
+        if not any((Cz, Mz, Sz, Vz, Kz, Wz)):
+            return [0.0, 0.0, 0.0]
+    if Qz == 0.0:
+        Qz = alg.EPS
+        if not any((Cz, Mz, Sz, Vz, Kz, Wz)):
+            return [0.0, 0.0, 0.0]
 
     # Break hue into Cartesian components
     h_rad = 0.0
@@ -424,12 +430,6 @@ class ZCAMJMh(LCh):
             return self.from_base(self.to_base(coords))
         coords[2] %= 360.0
         return coords
-
-    def is_achromatic(self, coords: Vector) -> bool | None:
-        """Check if color is achromatic."""
-
-        # Account for both positive and negative chroma
-        return coords[0] == 0 or abs(coords[1]) < self.achromatic_threshold
 
     def hue_name(self) -> str:
         """Hue name."""