+++ docs/src/markdown/about/changelog.md
@@ -2,11 +2,12 @@
 
 ## 5.0
 
+-   **NEW**: CAM16, CAM02, ZCAM, Hellwig, Luv, and LChuv will no longer force colors to black when lightness is zero
+    except when chroma/saturation/colorfulness is also zero. This allows out of gamut colors with lightness of zero
+    to properly be seen as out of gamut.
 -   **NEW**: CSS no longer requires HSL and HWB to return percentage form during serialization when using the non-legacy
     (no comma) syntax. For consistency, HSL and HWB will only return percentage form with non-legacy serialization  when
     `percent` is `True`. This matches all other color spaces in ColorAide.
--   **NEW**: Ray trace gamut mapping uses vector projection when correcting colors for better results and uses Oklab as
-    the default instead of OkLCh as using Oklab is a little faster.
 -   **NEW**: Add new `Luminant` mixin for color spaces which allows for internally targeting spaces that expose a
     lightness component.
 -   **NEW**: Rename `Regular` space mixin as `Prism` for a more apt description. `Regular` is still available but